fix: move type to separate file, delete unused translation
