+++ tests/integration/fixtures/test-pod-disruption-budgets.yaml
@@ -4,3 +4,8 @@ metadata:
   name: 'new-pdb'
 spec:
   minAvailable: 1
+status:
+  currentHealthy: 1
+  desiredHealthy: 1
+  disruptionsAllowed: 0
+  expectedPods: 1
